refactor: improve version detection and updating in build script
After speaking to some extension experts, I learned that it's actually
a bad practice to include a `package.json` file. Typically it's better
for this to get generated as part of the build process. This is because
including it results in duplication of metadata (e.g version), which
then needs to be updated in two places.

Instead of finding the version in `package.json`, we should be grabbing it
from the `manifest.json` file. This will allow extensions to stop
providing `package.json` and generate it as part of the build process.

Since there's already logic to find and update all `manifest.json`
files, I leveraged this to provide a solution that should work for any
extension, whether it provides package.json or not.